Removed UI test. Added 3DS extension file back.
