+++ Demo/UI Tests/Visa Checkout UI Tests/VisaCheckout_UITests.swift
@@ -11,52 +11,6 @@ class BraintreeVisaCheckout_UITests: XCTestCase {
         app.launchArguments.append("-TokenizationKey")
         app.launchArguments.append("-Integration:VisaCheckoutViewController")
         app.launch()
-        sleep(2)
-    }
-
-    func testVisaCheckout_whenNewCardAndBillingAddressAddedSuccessfully_returnToApp() {
-        let visaButton = app.buttons["visaCheckoutButton"]
-        let continueButton = app.buttons["CONTINUE"]
-        let expirationDate = UITestDateGenerator.sharedInstance.futureDate()
-
-        XCTAssertTrue(visaButton.waitForExistence(timeout: 60))
-        visaButton.forceTapElement()
-
-        XCTAssertTrue(app.webViews.buttons["New user?"].waitForExistence(timeout: 10))
-        app.webViews.buttons["New user?"].forceTapElement()
-
-        app.enterVisaCardDetailsWith(cardNumber: "4012000033330026", expirationDate: expirationDate)
- 
-        XCTAssertTrue(continueButton.waitForExistence(timeout: 60))
-        continueButton.forceTapElement()
-
-        app.enterBillingAddress(
-            firstName: "Joe",
-            lastName: "Doe",
-            addressLine1: "123 Main Street",
-            city: "Pleasanton",
-            state: "CA",
-            zipCode: "94533",
-            mobileNumber: "8642752333",
-            emailAddress: "joedoe@example.com"
-        )
-    
-        XCTAssertTrue(app.buttons["CONTINUE"].waitForExistence(timeout: 10))
-        continueButton.forceTapElement()
-        sleep(5)
-
-        handleNewUserScenerios(continueButton: continueButton)
-
-        XCTAssertTrue(app.buttons["CONTINUE AS GUEST"].waitForExistence(timeout: 10))
-        app.buttons["CONTINUE AS GUEST"].forceTapElement()
-
-        app.staticTexts["ADD DELIVERY ADDRESS"].waitForExistence(timeout: 30)
-        app.enterBillingAddress(addressLine1: "123 Main Street", state: "CA", zipCode: "94533")
-        
-        XCTAssertTrue(app.buttons["CONTINUE"].waitForExistence(timeout: 30))
-        continueButton.forceTapElement()
-
-        XCTAssertTrue(visaButton.waitForExistence(timeout: 60))
     }
 
     func testVisaCheckout_whenCanceled_returnToApp() {
@@ -72,23 +26,5 @@ class BraintreeVisaCheckout_UITests: XCTestCase {
         closeButtonCoordinate.doubleTap()
 
         XCTAssertTrue(visaButton.waitForExistence(timeout: 30))
-    }
-
-    // MARK: - Helper Function
-
-    /// This function handles the different flows that can occur for new users
-    private func handleNewUserScenerios(continueButton: XCUIElement) {
-        if app.staticTexts["VERIFY YOUR ADDRESS"].waitForExistence(timeout: 30) {
-            let recommendedButton = app.buttons["USE RECOMMENDED ADDRESS"]
-            XCTAssertTrue(recommendedButton.waitForExistence(timeout: 10))
-            recommendedButton.forceTapElement()
-        }
-
-        let welcomeBackText = app.descendants(matching: .any).matching(NSPredicate(format: "label == %@", "Welcome Back")).firstMatch
-        if welcomeBackText.waitForExistence(timeout: 30) {
-            let continueAsGuest = app.buttons["CONTINUE AS GUEST"]
-            XCTAssertTrue(continueAsGuest.waitForExistence(timeout: 10))
-            continueAsGuest.forceTapElement()
-        }
     }
 }
